challenges/Programmation/quick_response_code/requirements.txt
qrcode
pyzbar
pillow